Many changes because Pharo equates Symbols and Strings, allowing for sloppiness especially in test methods. Notably, affected synthesized column names in joins, too.
Changed set comparison to determine duplicates to use a bag, since nils cannot be added to hashed collections in GemStone/S.

GemStone/S and Pharo differ on which class is used when a Dictionary answers #keys. Pharo uses an Array, while GemStone/S uses a Set (as most other Smalltalk implementations do).

Changed a number of string encoded times to the format that GemStone/S handles (hh:mm:ss).

GemStone/S and Pharo use different classes for object not found in a collection, so that's been factored out from tests.

Pharo allow the argument to #and: to not be a block. Cleaned up those examples.

Added a number of extension methods that are only used by DataFrame, so far.

Worked around the deficiency in the current GemStone Zinc port which doesn't support reading single characters from external streams. Methods changed have been marked with a <HACK:...> pragma so they can be found and reverted once that deficiency has been addressed.

Specifically, the methods now follow a pattern like this:
	<HACK: 'Buffer the whole stream to work around lack of character decoding'>

	"Read data frame from a CSV file"

	| stream reader df |

	stream := aFileReference readStream.

	stream := ZnBufferedReadStream on: stream. "wrap with buffered stream to bypass https://github.com/GemTalk/FileSystemGs/issues/9"

	stream sizeBuffer: aFileReference size. "part of workaround for GemTalk/FileSystemGs#9"